Small cleanup and fixes
