+++ plugins/telemetry/telemetry.go
@@ -15,14 +15,12 @@
 package telemetry
 
 import (
-	"strconv"
 	"time"
 
 	"github.com/ligato/cn-infra/infra"
 	prom "github.com/ligato/cn-infra/rpc/prometheus"
 	"github.com/ligato/cn-infra/servicelabel"
 	"github.com/prometheus/client_golang/prometheus"
-	"github.com/prometheus/client_golang/prometheus/promhttp"
 
 	"github.com/ligato/vpp-agent/plugins/govppmux"
 	"github.com/ligato/vpp-agent/plugins/telemetry/vppcalls"
@@ -32,56 +30,8 @@ import (
 )
 
 const (
-	// Default update - period between metric updates
+	// Default update period between metric updates
 	defaultUpdatePeriod = time.Second * 30
-
-	// Registry path for telemetry metrics
-	registryPath = "/vpp"
-
-	// Metrics label used for agent label
-	agentLabel = "agent"
-)
-
-const (
-	// Runtime
-	runtimeThreadLabel   = "thread"
-	runtimeThreadIDLabel = "threadID"
-	runtimeItemLabel     = "item"
-
-	runtimeCallsMetric          = "calls"
-	runtimeVectorsMetric        = "vectors"
-	runtimeSuspendsMetric       = "suspends"
-	runtimeClocksMetric         = "clocks"
-	runtimeVectorsPerCallMetric = "vectors_per_call"
-
-	// Memory
-	memoryThreadLabel   = "thread"
-	memoryThreadIDLabel = "threadID"
-
-	memoryObjectsMetric   = "objects"
-	memoryUsedMetric      = "used"
-	memoryTotalMetric     = "total"
-	memoryFreeMetric      = "free"
-	memoryReclaimedMetric = "reclaimed"
-	memoryOverheadMetric  = "overhead"
-	memoryCapacityMetric  = "capacity"
-
-	// Buffers
-	buffersThreadIDLabel = "threadID"
-	buffersItemLabel     = "item"
-	buffersIndexLabel    = "index"
-
-	buffersSizeMetric     = "size"
-	buffersAllocMetric    = "alloc"
-	buffersFreeMetric     = "free"
-	buffersNumAllocMetric = "num_alloc"
-	buffersNumFreeMetric  = "num_free"
-
-	// Node counters
-	nodeCounterItemLabel   = "item"
-	nodeCounterReasonLabel = "reason"
-
-	nodeCounterCountMetric = "count"
 )
 
 // Plugin registers Telemetry Plugin
@@ -90,17 +40,7 @@ type Plugin struct {
 
 	handler vppcalls.TelemetryVppAPI
 
-	runtimeGaugeVecs map[string]*prometheus.GaugeVec
-	runtimeStats     map[string]*runtimeStats
-
-	memoryGaugeVecs map[string]*prometheus.GaugeVec
-	memoryStats     map[string]*memoryStats
-
-	buffersGaugeVecs map[string]*prometheus.GaugeVec
-	buffersStats     map[string]*buffersStats
-
-	nodeCounterGaugeVecs map[string]*prometheus.GaugeVec
-	nodeCounterStats     map[string]*nodeCounterStats
+	prometheusMetrics
 
 	// From config file
 	updatePeriod time.Duration
@@ -144,8 +84,10 @@ type nodeCounterStats struct {
 
 // Init initializes Telemetry Plugin
 func (p *Plugin) Init() error {
+	p.quit = make(chan struct{})
+
 	// Telemetry config file
-	config, err := p.getConfig()
+	config, err := p.loadConfig()
 	if err != nil {
 		return err
 	}
@@ -170,136 +112,8 @@ func (p *Plugin) Init() error {
 		p.updatePeriod = defaultUpdatePeriod
 	}
 
-	// Register '/vpp' registry path
-	err = p.Prometheus.NewRegistry(registryPath, promhttp.HandlerOpts{ErrorHandling: promhttp.ContinueOnError})
-	if err != nil {
+	if err := p.registerPrometheus(); err != nil {
 		return err
-	}
-
-	// Runtime metrics
-	p.runtimeGaugeVecs = make(map[string]*prometheus.GaugeVec)
-	p.runtimeStats = make(map[string]*runtimeStats)
-
-	for _, metric := range [][2]string{
-		{runtimeCallsMetric, "Number of calls"},
-		{runtimeVectorsMetric, "Number of vectors"},
-		{runtimeSuspendsMetric, "Number of suspends"},
-		{runtimeClocksMetric, "Number of clocks"},
-		{runtimeVectorsPerCallMetric, "Number of vectors per call"},
-	} {
-		name := metric[0]
-		p.runtimeGaugeVecs[name] = prometheus.NewGaugeVec(prometheus.GaugeOpts{
-			Namespace: "vpp",
-			Subsystem: "runtime",
-			Name:      name,
-			Help:      metric[1],
-			ConstLabels: prometheus.Labels{
-				agentLabel: p.ServiceLabel.GetAgentLabel(),
-			},
-		}, []string{runtimeItemLabel, runtimeThreadLabel, runtimeThreadIDLabel})
-
-	}
-
-	// register created vectors to prometheus
-	for name, metric := range p.runtimeGaugeVecs {
-		if err := p.Prometheus.Register(registryPath, metric); err != nil {
-			p.Log.Errorf("failed to register %v metric: %v", name, err)
-			return err
-		}
-	}
-
-	// Memory metrics
-	p.memoryGaugeVecs = make(map[string]*prometheus.GaugeVec)
-	p.memoryStats = make(map[string]*memoryStats)
-
-	for _, metric := range [][2]string{
-		{memoryObjectsMetric, "Number of objects"},
-		{memoryUsedMetric, "Used memory"},
-		{memoryTotalMetric, "Total memory"},
-		{memoryFreeMetric, "Free memory"},
-		{memoryReclaimedMetric, "Reclaimed memory"},
-		{memoryOverheadMetric, "Overhead"},
-		{memoryCapacityMetric, "Capacity"},
-	} {
-		name := metric[0]
-		p.memoryGaugeVecs[name] = prometheus.NewGaugeVec(prometheus.GaugeOpts{
-			Namespace: "vpp",
-			Subsystem: "memory",
-			Name:      name,
-			Help:      metric[1],
-			ConstLabels: prometheus.Labels{
-				agentLabel: p.ServiceLabel.GetAgentLabel(),
-			},
-		}, []string{memoryThreadLabel, memoryThreadIDLabel})
-
-	}
-
-	// register created vectors to prometheus
-	for name, metric := range p.memoryGaugeVecs {
-		if err := p.Prometheus.Register(registryPath, metric); err != nil {
-			p.Log.Errorf("failed to register %v metric: %v", name, err)
-			return err
-		}
-	}
-
-	// Buffers metrics
-	p.buffersGaugeVecs = make(map[string]*prometheus.GaugeVec)
-	p.buffersStats = make(map[string]*buffersStats)
-
-	for _, metric := range [][2]string{
-		{buffersSizeMetric, "Size of buffer"},
-		{buffersAllocMetric, "Allocated"},
-		{buffersFreeMetric, "Free"},
-		{buffersNumAllocMetric, "Number of allocated"},
-		{buffersNumFreeMetric, "Number of free"},
-	} {
-		name := metric[0]
-		p.buffersGaugeVecs[name] = prometheus.NewGaugeVec(prometheus.GaugeOpts{
-			Namespace: "vpp",
-			Subsystem: "buffers",
-			Name:      name,
-			Help:      metric[1],
-			ConstLabels: prometheus.Labels{
-				agentLabel: p.ServiceLabel.GetAgentLabel(),
-			},
-		}, []string{buffersThreadIDLabel, buffersItemLabel, buffersIndexLabel})
-
-	}
-
-	// register created vectors to prometheus
-	for name, metric := range p.buffersGaugeVecs {
-		if err := p.Prometheus.Register(registryPath, metric); err != nil {
-			p.Log.Errorf("failed to register %v metric: %v", name, err)
-			return err
-		}
-	}
-
-	// Node counters metrics
-	p.nodeCounterGaugeVecs = make(map[string]*prometheus.GaugeVec)
-	p.nodeCounterStats = make(map[string]*nodeCounterStats)
-
-	for _, metric := range [][2]string{
-		{nodeCounterCountMetric, "Count"},
-	} {
-		name := metric[0]
-		p.nodeCounterGaugeVecs[name] = prometheus.NewGaugeVec(prometheus.GaugeOpts{
-			Namespace: "vpp",
-			Subsystem: "node_counter",
-			Name:      name,
-			Help:      metric[1],
-			ConstLabels: prometheus.Labels{
-				agentLabel: p.ServiceLabel.GetAgentLabel(),
-			},
-		}, []string{nodeCounterItemLabel, nodeCounterReasonLabel})
-
-	}
-
-	// register created vectors to prometheus
-	for name, metric := range p.nodeCounterGaugeVecs {
-		if err := p.Prometheus.Register(registryPath, metric); err != nil {
-			p.Log.Errorf("failed to register %v metric: %v", name, err)
-			return err
-		}
 	}
 
 	return nil
@@ -311,8 +125,6 @@ func (p *Plugin) AfterInit() error {
 	if p.disabled {
 		return nil
 	}
-
-	p.quit = make(chan struct{})
 
 	go p.periodicUpdates()
 
@@ -343,7 +155,7 @@ Loop:
 		select {
 		// Delay period between updates
 		case <-time.After(p.updatePeriod):
-			p.updateData()
+			p.updatePrometheus()
 		// Plugin has stopped.
 		case <-p.quit:
 			break Loop
@@ -352,145 +164,4 @@ Loop:
 
 	// Close GoVPP channel
 	vppCh.Close()
-}
-
-func (p *Plugin) updateData() {
-	// Update runtime
-	runtimeInfo, err := p.handler.GetRuntimeInfo()
-	if err != nil {
-		p.Log.Errorf("Command failed: %v", err)
-	} else {
-		for _, thread := range runtimeInfo.Threads {
-			for _, item := range thread.Items {
-				stats, ok := p.runtimeStats[item.Name]
-				if !ok {
-					stats = &runtimeStats{
-						threadID:   thread.ID,
-						threadName: thread.Name,
-						itemName:   item.Name,
-						metrics:    map[string]prometheus.Gauge{},
-					}
-
-					// add gauges with corresponding labels into vectors
-					for k, vec := range p.runtimeGaugeVecs {
-						stats.metrics[k], err = vec.GetMetricWith(prometheus.Labels{
-							runtimeItemLabel:     item.Name,
-							runtimeThreadLabel:   thread.Name,
-							runtimeThreadIDLabel: strconv.Itoa(int(thread.ID)),
-						})
-						if err != nil {
-							p.Log.Error(err)
-						}
-					}
-				}
-
-				stats.metrics[runtimeCallsMetric].Set(float64(item.Calls))
-				stats.metrics[runtimeVectorsMetric].Set(float64(item.Vectors))
-				stats.metrics[runtimeSuspendsMetric].Set(float64(item.Suspends))
-				stats.metrics[runtimeClocksMetric].Set(item.Clocks)
-				stats.metrics[runtimeVectorsPerCallMetric].Set(item.VectorsPerCall)
-			}
-		}
-	}
-
-	// Update memory
-	memoryInfo, err := p.handler.GetMemory()
-	if err != nil {
-		p.Log.Errorf("Command failed: %v", err)
-	} else {
-		for _, thread := range memoryInfo.Threads {
-			stats, ok := p.memoryStats[thread.Name]
-			if !ok {
-				stats = &memoryStats{
-					threadName: thread.Name,
-					threadID:   thread.ID,
-					metrics:    map[string]prometheus.Gauge{},
-				}
-
-				// add gauges with corresponding labels into vectors
-				for k, vec := range p.memoryGaugeVecs {
-					stats.metrics[k], err = vec.GetMetricWith(prometheus.Labels{
-						memoryThreadLabel:   thread.Name,
-						memoryThreadIDLabel: strconv.Itoa(int(thread.ID)),
-					})
-					if err != nil {
-						p.Log.Error(err)
-					}
-				}
-			}
-
-			stats.metrics[memoryObjectsMetric].Set(float64(thread.Objects))
-			stats.metrics[memoryUsedMetric].Set(float64(thread.Used))
-			stats.metrics[memoryTotalMetric].Set(float64(thread.Total))
-			stats.metrics[memoryFreeMetric].Set(float64(thread.Free))
-			stats.metrics[memoryReclaimedMetric].Set(float64(thread.Reclaimed))
-			stats.metrics[memoryOverheadMetric].Set(float64(thread.Overhead))
-			stats.metrics[memoryCapacityMetric].Set(float64(thread.Capacity))
-		}
-	}
-
-	// Update buffers
-	buffersInfo, err := p.handler.GetBuffersInfo()
-	if err != nil {
-		p.Log.Errorf("Command failed: %v", err)
-	} else {
-		for _, item := range buffersInfo.Items {
-			stats, ok := p.buffersStats[item.Name]
-			if !ok {
-				stats = &buffersStats{
-					threadID:  item.ThreadID,
-					itemName:  item.Name,
-					itemIndex: item.Index,
-					metrics:   map[string]prometheus.Gauge{},
-				}
-
-				// add gauges with corresponding labels into vectors
-				for k, vec := range p.buffersGaugeVecs {
-					stats.metrics[k], err = vec.GetMetricWith(prometheus.Labels{
-						buffersThreadIDLabel: strconv.Itoa(int(item.ThreadID)),
-						buffersItemLabel:     item.Name,
-						buffersIndexLabel:    strconv.Itoa(int(item.Index)),
-					})
-					if err != nil {
-						p.Log.Error(err)
-					}
-				}
-			}
-
-			stats.metrics[buffersSizeMetric].Set(float64(item.Size))
-			stats.metrics[buffersAllocMetric].Set(float64(item.Alloc))
-			stats.metrics[buffersFreeMetric].Set(float64(item.Free))
-			stats.metrics[buffersNumAllocMetric].Set(float64(item.NumAlloc))
-			stats.metrics[buffersNumFreeMetric].Set(float64(item.NumFree))
-		}
-	}
-
-	// Update node counters
-	nodeCountersInfo, err := p.handler.GetNodeCounters()
-	if err != nil {
-		p.Log.Errorf("Command failed: %v", err)
-	} else {
-		for _, item := range nodeCountersInfo.Counters {
-			stats, ok := p.nodeCounterStats[item.Node]
-			if !ok {
-				stats = &nodeCounterStats{
-					itemName: item.Node,
-					metrics:  map[string]prometheus.Gauge{},
-				}
-
-				// add gauges with corresponding labels into vectors
-				for k, vec := range p.nodeCounterGaugeVecs {
-					stats.metrics[k], err = vec.GetMetricWith(prometheus.Labels{
-						nodeCounterItemLabel:   item.Node,
-						nodeCounterReasonLabel: item.Reason,
-					})
-					if err != nil {
-						p.Log.Error(err)
-					}
-				}
-			}
-
-			stats.metrics[nodeCounterCountMetric].Set(float64(item.Count))
-		}
-	}
 }
